// src/domain/entities/enums.ts
export enum CharacterStatus {
  Alive = "Alive",
  Dead = "Dead",
  Unknown = "unknown",
}

export enum CharacterGender {
  Female = "Female",
  Male = "Male",
  Genderless = "Genderless",
  Unknown = "unknown",
}

export enum CharacterType {
  All = "all",
  Starred = "starred",
  Others = "others",
}

export enum SpeciesType {
  All = "all",
  Human = "Human",
  Alien = "Alien",
}

export enum SortOrder {
  Asc = "asc",
  Desc = "desc",
}